+++ src/main/java/com/attica/athens/domain/agora/dto/AgoraMetrics.java
@@ -1,6 +1,6 @@
 package com.attica.athens.domain.agora.dto;
 
-public record AgoraMetrics(long agoraId, long membersCount, long chatCount) {
+public record AgoraMetrics(long agoraId, int membersCount, int chatCount) {
 }
